+++ CHANGELOG.md
@@ -12,10 +12,12 @@
 - Go to definition now goes to real file if such presented, or opens from the server
 - Basic syntax highlighting for CSP files, only as HTML
 - Added some snippets for class
-- View subclasses for current class, available in command palette
+- Go to Subclass for the current class, available in command palette
+- Go to Super class for the current class, available in command palette
 - Go To any class/method in the workspace including server (by Cmd+T/Ctrl+T)
 - some small fixes in the highlighting, and selecting words/variables
 - Intellisense. Show list of methods for ##class(SomeClass)
+- Go to macros definition
 
 ## [0.7.7]